move tests to config_tests to eliminate config_api solana_runtime dependency
