+++ programs/config_tests/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "solana-config-tests"
+version = "0.18.0-pre1"
+description = "Solana config api tests"
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+edition = "2018"
+
+[dependencies]
+bincode = "1.1.4"
+log = "0.4.8"
+serde = "1.0.98"
+serde_derive = "1.0.98"
+solana-logger = { path = "../../logger", version = "0.18.0-pre1" }
+solana-sdk = { path = "../../sdk", version = "0.18.0-pre1" }
+solana-config-api = { path = "../config_api", version = "0.18.0-pre1" }
+solana-config-program = { path = "../config_program", version = "0.18.0-pre1" }
+
+
+[dev-dependencies]
+solana-runtime = { path = "../../runtime", version = "0.18.0-pre1" }
+assert_matches = "1.3.0"
